Move `util_bloqs.py` to `bookkeeping/` (#999)
* Move util_bloqs.py to bookkeeping/

* Fix formatting and upate remaining references
