+++ qualtran/bloqs/arithmetic/addition.py
@@ -55,8 +55,8 @@ from qualtran import (
     Soquet,
     SoquetT,
 )
-from qualtran.bloqs import util_bloqs
 from qualtran.bloqs.basic_gates import CNOT, XGate
+from qualtran.bloqs.bookkeeping import util_bloqs
 from qualtran.bloqs.mcmt.and_bloq import And
 from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import MultiControlX
 from qualtran.cirq_interop import decompose_from_cirq_style_method
