+++ qualtran/bloqs/bookkeeping/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright 2023 Google LLC
+#  Copyright 2024 Google LLC
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,15 +11,13 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import qualtran.testing as qlt_testing
-from qualtran import QAny
-from qualtran.bloqs.basic_gates.swap import Swap
-from qualtran.bloqs.bookkeeping import Allocate, Free
-from qualtran.bloqs.for_testing.interior_alloc import InteriorAlloc
-
-
-def test_interior_alloc():
-    ia = InteriorAlloc(10)
-    qlt_testing.assert_valid_bloq_decomposition(ia)
-    g, counts = ia.call_graph(max_depth=1)
-    assert counts == {Allocate(QAny(10)): 1, Free(QAny(10)): 1, Swap(10): 2}
+from qualtran.bloqs.bookkeeping.util_bloqs import (
+    Allocate,
+    ArbitraryClifford,
+    Cast,
+    Free,
+    Join,
+    Partition,
+    Power,
+    Split,
+)
