+++ qualtran/bloqs/factoring/mod_mul_test.py
@@ -21,9 +21,9 @@ import sympy
 
 import qualtran.testing as qlt_testing
 from qualtran import Bloq
+from qualtran.bloqs.bookkeeping import Allocate, Free
 from qualtran.bloqs.factoring.mod_mul import _modmul, _modmul_symb, CtrlModMul, MontgomeryModDbl
 from qualtran.bloqs.mod_arithmetic import CtrlScaleModAdd
-from qualtran.bloqs.util_bloqs import Allocate, Free
 from qualtran.drawing import Text
 from qualtran.resource_counting import SympySymbolAllocator
 from qualtran.testing import assert_valid_bloq_decomposition
