+++ qualtran/bloqs/multiplexers/unary_iteration_bloq_test.py
@@ -22,8 +22,8 @@ import pytest
 from qualtran import BoundedQUInt, QAny, Register, Signature
 from qualtran._infra.gate_with_registers import get_named_qubits, total_bits
 from qualtran.bloqs.basic_gates import CNOT
+from qualtran.bloqs.bookkeeping import Join, Split
 from qualtran.bloqs.multiplexers.unary_iteration_bloq import unary_iteration, UnaryIterationGate
-from qualtran.bloqs.util_bloqs import Join, Split
 from qualtran.cirq_interop.bit_tools import iter_bits
 from qualtran.cirq_interop.testing import assert_circuit_inp_out_cirqsim, GateHelper
 from qualtran.resource_counting.generalizers import cirq_to_bloqs
